Keep lots of refactoring...making way for more
